Efetuado ajustes e implementação de logica

+++ src/main/java/com/finance/transaction/model/Transfer.java
@@ -11,7 +11,7 @@ public class Transfer {
 
     String originAccount;
     String destinationAccount;
-    Integer value;
+    double value;
     LocalDateTime dateTransfer;
     LocalDateTime dateScheduling;
